LayoutTests/media/video-currentTime-expected.txt

EXPECTED (video.currentTime == '0') OK
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
EVENT(play)
EXPECTED (video.currentTime > '0') OK
END OF TEST
